refactor: move hooks to separate folder
