+++ src/pages/home/report/ReportActionCompose/SendButton.tsx
@@ -5,12 +5,12 @@ import Icon from '@components/Icon';
 import * as Expensicons from '@components/Icon/Expensicons';
 import PressableWithFeedback from '@components/Pressable/PressableWithFeedback';
 import Tooltip from '@components/Tooltip';
+import useFreezeId from '@hooks/useFreezeId';
 import useLocalize from '@hooks/useLocalize';
 import useResponsiveLayout from '@hooks/useResponsiveLayout';
 import useTheme from '@hooks/useTheme';
 import useThemeStyles from '@hooks/useThemeStyles';
 import CONST from '@src/CONST';
-import useFreezeId from './useFreezeId';
 
 type SendButtonProps = {
     /** Whether the button is disabled */
